refactor(memo): update main class and configuration for memo microservice
