+++ apps/backend/memo/infra/build.gradle.kts
@@ -23,7 +23,7 @@ dependencies {
 }
 
 application {
-    mainClass = "top.sunbath.api.auth.Main"
+    mainClass = "top.sunbath.api.memo.Main"
 }
 
 tasks.withType<Test> {
