+++ apps/backend/memo/infra/src/main/java/top/sunbath/api/memo/Main.java
@@ -1,11 +1,11 @@
-package top.sunbath.api.auth;
+package top.sunbath.api.memo;
 
 import software.amazon.awscdk.App;
 
 public class Main {
     public static void main(final String[] args) {
         App app = new App();
-        new AppStack(app, "AuthStack");
+        new AppStack(app, "MemoStack");
         app.synth();
     }
-} 
+}
