+++ apps/backend/memo/infra/src/test/java/top/sunbath/api/auth/AppStackTest.java
@@ -1,8 +1,10 @@
-package top.sunbath.api.auth;
+package top.sunbath.api.memo;
 
 import org.junit.jupiter.api.Test;
 import software.amazon.awscdk.App;
 import software.amazon.awscdk.assertions.Template;
+import top.sunbath.api.memo.AppStack;
+
 import java.io.File;
 import java.util.Collections;
 
@@ -11,9 +13,10 @@ class AppStackTest {
     @Test
     void testAppStack() {
         if (new File(AppStack.functionPath()).exists()) {
-            AppStack stack = new AppStack(new App(), "TestAuthAppStack");
+            AppStack stack = new AppStack(new App(), "TestMemoAppStack");
             Template template = Template.fromStack(stack);
-            template.hasResourceProperties("AWS::Lambda::Function", Collections.singletonMap("Handler", "io.micronaut.function.aws.proxy.payload2.APIGatewayV2HTTPEventFunction"));
+            template.hasResourceProperties("AWS::Lambda::Function", Collections.singletonMap("Handler",
+                    "io.micronaut.function.aws.proxy.payload2.APIGatewayV2HTTPEventFunction"));
         }
     }
-} 
+}
